# backend/src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "MessageDefault": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "sentWhen": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DefaultChatResponse": {
        "type": "object",
        "properties": {
          "patient": {
            "type": "string"
          },
          "attendant": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDefault"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/chats/{id}": {
      "get": {
        "tags": ["chats"],
        "summary": "Get chat by id",
        "operationId": "getChatById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultChatResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/chats/{id}/accept": {
      "post": {
        "tags": ["chats"],
        "summary": "Accept chat",
        "operationId": "acceptChat",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}